refactor: focusTrap customHook으로 분리
